# backend/srp-base/openapi/api.yaml
openapi: 3.0.3
info:
  title: SunnyRP Base API
  version: 0.1.0
  description: |
    This API powers the SunnyRP backend.  All persistence and
    authoritative state resides behind these routes.  Consumers (e.g.
    FiveM Lua) must authenticate with X-API-Token on every call.
servers:
  - url: http://localhost:3010
components:
  securitySchemes:
    ApiToken:
      type: apiKey
      in: header
      name: X-API-Token
security:
  - ApiToken: []
paths:
  /v1/healthz:
    get:
      summary: Liveness probe
      responses:
        '200':
          description: Service is alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
  /v1/ready:
    get:
      summary: Readiness probe
      responses:
        '200':
          description: Service is ready
        '503':
          description: Dependencies unavailable
  /v1/config/live:
    get:
      summary: Retrieve live config
      responses:
        '200':
          description: Current configuration
    post:
      summary: Update live config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                features:
                  type: object
                settings:
                  type: object
      responses:
        '200':
          description: Updated config
  /v1/users/exists:
    get:
      summary: Check if a user exists
      parameters:
        - in: query
          name: hex_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Existence result
  /v1/users:
    post:
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - hex_id
                - name
              properties:
                hex_id:
                  type: string
                name:
                  type: string
                identifiers:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                      value:
                        type: string
      responses:
        '200':
          description: Created user
  /v1/users/{hex_id}:
    get:
      summary: Get a user
      parameters:
        - name: hex_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User object
        '404':
          description: Not found
  /v1/characters:
    get:
      summary: List characters for an owner
      parameters:
        - name: owner_hex
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of characters
    post:
      summary: Create a character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - owner_hex
                - first_name
                - last_name
              properties:
                owner_hex:
                  type: string
                first_name:
                  type: string
                last_name:
                  type: string
      responses:
        '200':
          description: Created character
  /v1/characters/{id}:
    get:
      summary: Get a character by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Character object
        '404':
          description: Not found
    patch:
      summary: Update a character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated character
  /v1/permissions/{playerId}:
    get:
      summary: List permissions for a player
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of scopes
  /v1/permissions/grant:
    post:
      summary: Grant a permission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - playerId
                - scope
              properties:
                playerId:
                  type: string
                scope:
                  type: string
      responses:
        '200':
          description: Updated scopes
  /v1/permissions/revoke:
    post:
      summary: Revoke a permission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - playerId
                - scope
              properties:
                playerId:
                  type: string
                scope:
                  type: string
      responses:
        '200':
          description: Updated scopes
  /v1/players/link:
    post:
      summary: Link a player by identifiers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - identifiers
              properties:
                identifiers:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                      value:
                        type: string
      responses:
        '200':
          description: Link result with user and scopes
  /v1/outbox/enqueue:
    post:
      summary: Enqueue an outbox event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - topic
                - payload
              properties:
                topic:
                  type: string
                payload:
                  type: object
      responses:
        '200':
          description: Created outbox event
  /v1/admin/ban:
    post:
      summary: Ban a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - playerId
                - reason
              properties:
                playerId:
                  type: string
                reason:
                  type: string
                until:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Player banned